requirements.txt
streamlit
pyaudio
numpy
openai-whisper
